+++ core/app_client.go
@@ -615,3 +615,17 @@ func (a appClient) requireLeaderboardAdmin(leaderboardID string, orgID string, a
 func newAppClient(app *Application) appClient {
 	return appClient{app: app}
 }
+
+// PrepareScoresForResponse prepares multiple scores for API response based on feature flag
+func (app *Application) PrepareScoresForResponse(scores []model.Score) {
+	if !app.useExternalUserID {
+		return
+	}
+
+	for i := range scores {
+		if scores[i].ExternalUserID != "" {
+			// When flag is enabled and we have an AmgUUID, use it in the external_profile_id field
+			scores[i].ExternalProfileID = scores[i].ExternalUserID
+		}
+	}
+}
