+++ core/app_shared.go
@@ -174,14 +174,9 @@ func (a appShared) createScore(orgID string, appID string, userID string, extern
 		a.updateScore(&score, surveyResponses[i], nil)
 	}
 
-	// Sync AmgUUID from Core BB before saving the score
-	// This populates the external_user_id field if we have a mastodon_id
-	// Note: SyncAmgUUIDForScore handles errors gracefully (returns nil on errors, logs internally)
-	// Check if amg_uuid is present in externalIDs
+	// Set AmgUUID from externalIDs if present
 	if amgUUID, ok := externalIDs["amg_uuid"]; ok && amgUUID != "" {
 		score.ExternalUserID = amgUUID
-	} else {
-		a.app.SyncAmgUUIDForScore(&score, externalIDs)
 	}
 
 	if apply {
